+++ CHANGELOG.md
@@ -6,6 +6,7 @@ FEATURES:
 
 * Added Authentication supports via OAuth2.0 to `Security` pillar
 * HTTP utils includes set and get cookies features.
+* Separate capability to generate Access Token for Client credentials from authorization code flow grant type
 
 
 ## v0.1.0 05/10/2023